--- src/hello.c
@@ -1,5 +1,0 @@
-#include "hello.h"
-
-char* hello() {
-    return "hello";
-}
