--- src/hello.h
@@ -1,4 +1,0 @@
-#pragma once
-#include <string.h>
-
-char* hello();
