--- test/test_hello.c
@@ -1,9 +1,0 @@
-#include <criterion/criterion.h>
-#include <criterion/new/assert.h>
-
-#include "../src/hello.h"
-
-Test(strip, empty_string) {
-  cr_assert(eq(str, hello(), "hello"));
-}
-
